requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.0.1
pyttsx3==2.90
